feat: SystemUI 관련 코드를 BaseActivity로 이동
