+++ modules/presentation/src/main/java/dev/yjyoon/kwnotice/presentation/ui/splash/SplashActivity.kt
@@ -1,51 +1,24 @@
 package dev.yjyoon.kwnotice.presentation.ui.splash
 
 import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.runtime.SideEffect
 import androidx.lifecycle.lifecycleScope
-import com.google.accompanist.systemuicontroller.SystemUiController
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import dagger.hilt.android.AndroidEntryPoint
+import dev.yjyoon.kwnotice.presentation.ui.base.BaseActivity
 import dev.yjyoon.kwnotice.presentation.ui.main.MainActivity
-import dev.yjyoon.kwnotice.presentation.ui.theme.KWNoticeDarkColors
-import dev.yjyoon.kwnotice.presentation.ui.theme.KWNoticeLightColors
-import dev.yjyoon.kwnotice.presentation.ui.theme.KWNoticeTheme
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
-class SplashActivity : ComponentActivity() {
+class SplashActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        setContent()
+        setScreen { SplashScreen() }
 
         lifecycleScope.launch {
             delay(SPLASH_TIME_MILLIS)
             startMainActivity()
-        }
-    }
-
-    private fun setContent() = setContent {
-        val systemUiController: SystemUiController = rememberSystemUiController()
-        val isDarkMode = isSystemInDarkTheme()
-        SideEffect {
-            systemUiController.setSystemBarsColor(
-                color = if (isDarkMode) {
-                    KWNoticeDarkColors.background
-                } else {
-                    KWNoticeLightColors.background
-                },
-                darkIcons = !isDarkMode
-            )
-        }
-
-        KWNoticeTheme {
-            SplashScreen()
         }
     }
